Create the POST api/auth/signup endpoint that accepts data in the JSON format: {    "name": "<String value, not empty>",    "lastname": "<String value, not empty>",    "email": "<String value, not empty>",    "password": "<String value, not empty>" }
It should return a response in the JSON format (without the password field):

{
   "name": "<String value>",
   "lastname": "<String value>",
   "email": "<String value>"
}

If the status is HTTP OK (200), then all fields are correct. If it's HTTP Bad Request (400), then something is wrong. Our service must accept only corporate emails that end with @acme.com. In this stage, we do not check the authentication, so the password field may contain anything (but not empty).