Part 2 - 3 Objectives Add the Spring security to your project and configure the HTTP basic authentication;
For storing users and passwords, add a JDBC implementation of UserDetailsService with an H2 database;

Change the POST api/auth/signup endpoint. It must be available to unauthorized users for registration and accepts data in the JSON format:

{
   "name": "<String value, not empty>",
   "lastname": "<String value, not empty>",
   "email": "<String value, not empty>",
   "password": "<String value, not empty>"
}
If OK, provide the HTTP OK status (200) and the following body:

{
   "id": "<Long value, not empty>",
   "name": "<String value, not empty>",
   "lastname": "<String value, not empty>",
   "email": "<String value, not empty>"
}
As a unique login for authentication, take the value from the email field. The value of the email field must be case insensitive. Id is a unique identifier that the service assigns to the user. If an email is occupied, the service should respond as shown below. The rest of the error messages are the same as in the previous stage:

{
    "timestamp": "data",
    "status": 400,
    "error": "Bad Request",
    "message": "User exist!",
    "path": "/api/auth/signup"
}
Add the GET api/empl/payment/ endpoint that allows for testing the authentication. It should be available only to authenticated users and return a response in the JSON format representing the user who has sent the request:

{
   "id": "<Long value, not empty>",
   "name": "<String value, not empty>",
   "lastname": "<String value, not empty>",
   "email": "<String value, not empty>"
}
The email field must contain the user's login who has sent the request. Error message for the non-authenticated or wrong user should have the 401 (Unauthorized) status.

Examples
Example 1: a POST request for api/auth/signup with the correct user

Request body:

{
   "name": "John",
   "lastname": "Doe",
   "email": "JohnDoe@acme.com",
   "password": "secret"
}
Response: 200 OK

Response body:

{
   "id": 1,
   "name": "John",
   "lastname": "Doe",
   "email": "JohnDoe@acme.com"
}
Example 2: a POST request for api/auth/signup with the occupied email

Request body:

{
   "name": "John",
   "lastname": "Doe",
   "email": "johndoe@acme.com",
   "password": "secret"
}
Response: 400 Bad Request

Response body:

{
    "timestamp": "<data>",
    "status": 400,
    "error": "Bad Request",
    "message": "User exist!",
    "path": "/api/auth/signup"
}
Example 3: a POST request for api/auth/signup with the wrong format of the user JSON

Request body:

{
   "lastname": "Doe",
   "email": "johndoe@acme.com",
   "password": "secret"
}
Response: 400 Bad Request

Response body:

{
   "timestamp": "<date>",
   "status": 400,
   "error": "Bad Request",
   "path": "/api/auth/signup"
}
Example 4: a GET request for /api/empl/payment with the correct authentication, username = johndoe@acme.com, password = secret

Response: 200 OK

Response body:

{
    "id": 1,
    "name": "John",
    "lastname": "Doe",
    "email": "johndoe@acme.com"
}
Example 5: a GET request for /api/empl/payment with the correct authentication; username = JohnDoe@acme.com, password = secret

Response: 200 OK

Response body:

{
    "id": 1,
    "name": "John",
    "lastname": "Doe",
    "email": "johndoe@acme.com"
}
Example 6: a GET request for /api/empl/payment with the wrong authentication; username = johndoe@acme.com, password = no_secret

Response body:

{
    "timestamp": "<data>",
    "status": 401,
    "error": "Unauthorized",
    "message": "",
    "path": "/api/empl/payment"
}

PART 3
Here they are:

Verify that user passwords contain at least 12 characters;
Verify that users can change their passwords. Changing the password requires the current and a new password;
Verify that the passwords submitted during a registration, login, and password change are checked against a set of breached passwords. If the password is breached, the application must require users to set a new non-breached password.
Verify that passwords are stored in a form that is resistant to offline attacks. Passwords must be salted and hashed using an approved one-way key derivation or a password hashing function;
If you use bcrypt, the work factor must be as large as the verification server performance will allow. Usually, at least 13.
At this stage, your need to implement the API for changing a user's password.

POST api/auth/changepass is designed to change the password of a user who was successfully authenticated during a request. It must accept data in the JSON format and update a password to the one specified in the new_password field. The old and new passwords must be different. A new password must meet the requirements listed above.
{"PasswordForJanuary", "PasswordForFebruary", "PasswordForMarch", "PasswordForApril",
 "PasswordForMay", "PasswordForJune", "PasswordForJuly", "PasswordForAugust",
 "PasswordForSeptember", "PasswordForOctober", "PasswordForNovember", "PasswordForDecember"}

 Examples
 Example 1: a POST request for api/auth/signup

 Request body:

 {
    "name": "John",
    "lastname": "Doe",
    "email": "johndoe@acme.com",
    "password": "secret"
 }
 Response: 400 Bad Request

 Response body:

 {
     "timestamp": "<data>",
     "status": 400,
     "error": "Bad Request",
     "message": "The password length must be at least 12 chars!",
     "path": "/api/auth/signup"
 }
 Example 2: a POST request for api/auth/signup

 Request body:

 {
    "name": "John",
    "lastname": "Doe",
    "email": "johndoe@acme.com",
    "password": "PasswordForJune"
 }
 Response: 400 Bad Request

 Response body:

 {
     "timestamp": "<data>",
     "status": 400,
     "error": "Bad Request",
     "message": "The password is in the hacker's database!",
     "path": "/api/auth/signup"
 }
 Example 3: a POST request for api/auth/changepass with the correct authentication; username = johndoe@acme.com, password = B3Fagws6zcBa

 Request body:

 {
    "new_password": "bZPGqH7fTJWW"
 }
 Response: 200 OK

 Response body:

 {
     "email": "johndoe@acme.com",
     "status": "The password has been updated successfully"
 }
 Example 4: a POST request for api/auth/changepass with the correct authentication; username = johndoe@acme.com, password = bZPGqH7fTJWW

 Request body:

 {
    "new_password": "bZPGqH7fTJWW"
 }
 Response: 400 Bad Request

 Response body:

 {
     "timestamp": "<date>",
     "status": 400,
     "error": "Bad Request",
     "message": "The passwords must be different!",
     "path": "/api/auth/changepass"
 }